+++ components/chat/reporting.tsx
@@ -5,6 +5,7 @@ import React, { Dispatch, SetStateAction, useEffect, useState, useContext } from
 import { Input } from "@/components/ui/input";
 import Image from "next/image";
 import { Button } from "@/components/ui/button";
+import {getGroups} from "@/utils/client_side/chatUtils"
 import {
     Select,
     SelectGroup,
@@ -22,34 +23,6 @@ import {doc, setDoc, Timestamp} from "firebase/firestore"
 import * as z from "zod"
 import axios from 'axios'
 
-async function getGroups(user: User): Promise<any> {
-    function getPeople(data: any) {
-        let collect: { [key: string]: any } = {};
-        data.forEach((chat: { title: any; people: { person: { username: any; }; }[]; }) => {
-            const title: string = chat.title;
-            const people: any = [];
-            chat.people.forEach((something: { person: { username: any; }; }) => {
-                people.push(something.person.username)
-            })
-            collect[title] = people
-        })
-        return collect
-    }
-    let groups: Object = {};
-    axios({
-        method: 'get',
-        url: 'https://api.chatengine.io/chats/',
-        headers: {
-            'project-id': process.env.NEXT_PUBLIC_CHAT_PROJECT as string,
-            'user-name': user?.firstName + " " + user?.lastName,
-            'user-secret': user?.id as string
-        }
-    })
-        .then((response) => { groups = getPeople(response.data) })
-        .catch((error) => { throw new Error(error) });
-    return groups;
-}
-
 export default function ReportForm() {
     const [groups, setGroups] = useState([""])
     const [reporting, setReporting] = useState(false)
@@ -64,8 +37,7 @@ export default function ReportForm() {
             try {
                 if (user) {
                     const result = await getGroups(user);
-                    setGroups(result);
-                    console.log(groups)
+                    console.log(result)
                 }
             } catch (error) {
                 console.error("Error fetching groups:", error);
